pyproject.toml
[project]
name = "sprint-4-iot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "oracledb>=3.4.0",
    "paho-mqtt>=2.1.0",
    "python-decouple>=3.8",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
    "streamlit-autorefresh>=1.0.1",
]
